# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API_GET_POST_UPDATE _DELETE_DATA",
        "description": "API PROJECT MODEL CHANGE RECORD",
        "version": "1.0.0"
    },
    "paths": {
        "/get/won": {
            "get": {
                "tags": [
                    "Work Order"
                ],
                "summary": "Get work order to model change",
                "operationId": "8404e63177adb559cbbef2d892ce995f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/modelname": {
            "get": {
                "tags": [
                    "Model Name"
                ],
                "summary": "Get data model name to model change",
                "operationId": "8bcd897143c43d8965ac38af44be5c45",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/line": {
            "get": {
                "tags": [
                    "Line Model Change"
                ],
                "summary": "Get data line to model change",
                "operationId": "9896b096429302b1ace9ece7523e133c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/checkmodel": {
            "get": {
                "tags": [
                    "Check Model"
                ],
                "summary": "Get work order and model name check match",
                "operationId": "876cf6ba1e227957c35fecb810c9435f",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Work Order",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/get/count/opr": {
            "get": {
                "tags": [
                    "Count of Record Form"
                ],
                "summary": "Get count of record",
                "operationId": "47b5aa7c49cc69e15592de823cb15517",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/record/opr": {
            "get": {
                "tags": [
                    "Get Record Form"
                ],
                "summary": "Get record data",
                "operationId": "ef2c541134e629d6d8a1bd19de749942",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/oprform": {
            "get": {
                "tags": [
                    "Get Record Form"
                ],
                "summary": "Get record data status = 1",
                "operationId": "15548af3bb62d3ee90ef77caac73f38d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/get/alldata": {
            "get": {
                "tags": [
                    "Get Record Form"
                ],
                "summary": "Get All Record Data",
                "operationId": "970b6098d350b87cafeaa568554a5209",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
...
}